fix return alue v0

+++ pack/pack.go
@@ -2,5 +2,5 @@ package pack
 
 // GetIntFromInt retuns the input
 func GetIntFromInt(i int) int {
-	return float64(i)
+	return i
 }
